webkit native/base/content update
web native/content update

+++ Sources/Elastos/Frameworks/Droid/Base/Core/src/elastos/droid/webkit/webview/chromium/native/android_webview/AwSettings.cpp
@@ -267,7 +267,7 @@ void AwSettings::EventHandler::RunOnUiThreadBlockingAndLocked(
                 ECode ecode = mOwner->mAwSettingsLock.Wait();
                 if (FAILED(ecode))
                 {
-                    if ((unsigned)ecode == (ECode)E_INTERRUPTED_EXCEPTION)
+                    if (ecode == (ECode)E_INTERRUPTED_EXCEPTION)
                         mSynchronizationPending = false;
                     break;
                 }
